+++ api/client.go
@@ -5,10 +5,11 @@ import (
 )
 
 type Client interface {
-	ListIPs(ctx context.Context) ([]IP, error)
+	ListIPs(context.Context) ([]IP, error)
 	GetIP(ctx context.Context, id string) (IP, error)
 	AddIP(ctx context.Context, ip string, params AddIPParams) (IP, error)
+	UpdateIP(ctx context.Context, id string, params UpdateIPParams) (IP, error)
 	RemoveIP(ctx context.Context, id string) error
 	Failover(ctx context.Context, id string) error
-	Version(ctx context.Context) (string, error)
+	Version(context.Context) (string, error)
 }