+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 * Remove `/try-get-lock` route and `try-get-lock` command in favor of `failover`
 * Destroying an IP is now synchronous
 * Remove `KeepaliveInterval` from the IP model
+* Add a way to update IP healthchecks
+* Bad request body key is `error` instead of `msg`
+* Not found body is `{"resource": "IP", "error" : "not found"}`
 
 ## [2020-11-20] 1.9.3